refactor: 이넘 constants 로 이동
