+++ app/constants.py
@@ -1,4 +1,5 @@
 import datetime
+from enum import Enum
 from app.config import settings
 
 
@@ -59,3 +60,23 @@ HELP_TEXT = f"""
 
 
 CONTENTS_PER_PAGE = 20
+
+
+class ContentCategoryEnum(str, Enum):
+    UDEMY = "유데미 후기"
+    PROJECT = "프로젝트"
+    TECH = "기술 & 언어"
+    CULTURE = "조직 & 문화"
+    JOB = "취준 & 이직"
+    DAILY = "일상 & 생각"
+    ETC = "기타"
+
+
+class ArchiveMessageSortEnum(str, Enum):
+    DT = "dt"
+
+
+class ContentSortEnum(str, Enum):
+    DT = "dt"
+    RELEVANCE = "relevance"
+    # LIKE = "like" # TODO: 추후 추가하기
